Loosened validation of submodule URL.
These aren't strictly "URLs" in the web sense. They may be a variety of
formats and don't strictly conform to HTTP-style URLs, like SSH style
URLs. This change aligns the validation with Git's own URL validation.
While the validation doesn't necessarily gaurantee that the URL is
actually valid, calling the checkout action (which would be necessarily
to use this action anyway) would fail before this if it couldn't clone a
submodule due to a bad URL.

- Switched from `.url()` to `.regex()` string validator using the regex
  present in url.c[1] in Git's source code.
- Added unit test and test fixture to check compatibility with SSH-style
  URLs.

[1]:
https://github.com/git/git/blob/77bd3ea9f54f1584147b594abc04c26ca516d987/url.c#L6-L12