Improved remote name extraction.
Since the action no longer accepts HTTP-style URLs, we need to account
for more cases. This commit adds a function that attempts a "best guess"
remote name extraction. It looks for specific characters in the URL that
probably demarcate the remote name. It's not a perfect solution, but it
works well in the vast majority of cases.

- Added `getRemoteName` function to get remote name.
- Added unit tests with a bunch of example cases that cover a gamut of
  different URL formats. The test cases were taken from this Stack
  Overflow post: https://stackoverflow.com/questions/31801271/what-are-the-supported-git-url-formats