fix: support multiple URL protocols (#76)
* Loosened validation of submodule URL.

These aren't strictly "URLs" in the web sense. They may be a variety of
formats and don't strictly conform to HTTP-style URLs, like SSH style
URLs. This change aligns the validation with Git's own URL validation.
While the validation doesn't necessarily gaurantee that the URL is
actually valid, calling the checkout action (which would be necessarily
to use this action anyway) would fail before this if it couldn't clone a
submodule due to a bad URL.

- Switched from `.url()` to `.regex()` string validator using the regex
  present in url.c[1] in Git's source code.
- Added unit test and test fixture to check compatibility with SSH-style
  URLs.

[1]:
https://github.com/git/git/blob/77bd3ea9f54f1584147b594abc04c26ca516d987/url.c#L6-L12

* Improved remote name extraction.

Since the action no longer accepts HTTP-style URLs, we need to account
for more cases. This commit adds a function that attempts a "best guess"
remote name extraction. It looks for specific characters in the URL that
probably demarcate the remote name. It's not a perfect solution, but it
works well in the vast majority of cases.

- Added `getRemoteName` function to get remote name.
- Added unit tests with a bunch of example cases that cover a gamut of
  different URL formats. The test cases were taken from this Stack
  Overflow post: https://stackoverflow.com/questions/31801271/what-are-the-supported-git-url-formats

* chore: delete helper function & rename test

---------

Co-authored-by: sgoudham <sgoudham@gmail.com>